第十次commit
封装程序，因为单元测试需要封装的程序
